Rename files to rtcSlot.ts, rtcSlot.test.ts
