+++ packages/react-sdk/src/model/rtcSlot.ts
